Redrew all UML diagrams using PlantUML for easier maintenance.
This allow us to track changes in the UML diagrams using git,
and thanks to the use of include files, consistent class
definitions across UML diagrams.

This further make it possible to generate the UML diagram
images automatically from the source files using a command
line free software tool, avoiding the need to accept a non
free software license to be able to correct the UML diagrams.

Fixes #76