Added text2uml tool to generate PlantUML diagram from the text in chapter 7.
This extracts class names, types, inheritence, relations and attributes
from the class definitions in chapter 7, and generate iuml and puml
files to get a UML model of the entire data structure.  It allow us
to more easily find bugs in the model, and to maintain the PlantUML
include files (see issue #76) using the textual description.

The relation links need to be verified, as there were some of them
I was unsure how to best model in PlantUML.