Use only entities from this specification in UML notation example.
The original UML notation example introduce entities that are not part of
this specification, like Foto, Fil, Albumdel etc.  This is confusing for
the reader when trying to figure out what these classes have to to with
this specification.  It is better to show the notation used using entities
that are part of this specification.  Rewrote the UML diagram to use
Arkivenhet, Mappe, Registrering, Journalpost, Avskrivning and related
code lists and relations.