Link to Unicode lists of Control and Space Separator characters
This make it easier in the future to check if new characters have been added
to the list.  It also make the standard text more dependent on
an external source, which can change at any time.  It was the
best source I could find listing characters in the classes in question.