Extend and explain in more details how to upload large files.
Introduce new bulkgrense system value to allow clients to learn exactly how small
small files are.  Specify how to reject large files using the small method.  Explain
the format of Range and Content-Range headers.  Made it clear that the upload need
to be done in sequence (and not out of order) by specifying that the Location header
returned after each bulk upload is to be used in the next bulk upload.

Drop the Google Drive reference.  It do not really explain much, the URL is broken
and this specification should describe its own protocol as it is not identical to
the Google Drive protocol.

Fixes #313.